XFAIL malloc_zone.cpp for darwin/lsan
The malloc_zone.cpp test currently fails on Darwin hosts, in SanitizerCommon
tests with lsan enabled.

Need to XFAIL this test to buy time to investigate this failure. Also
we're trying to bring the number of test failing on Darwin bots to 0, to
get clearer signal of any new failures.

rdar://145873843